Moving Axe HTML report CSS and JS to their own files
