+++ Lombiq.Tests.UI/AccessibilityChecking/AxeHtmlReport.cs
@@ -10,9 +10,10 @@ using System.Globalization;
 using System.IO;
 using System.Linq;
 using System.Net;
+using System.Reflection;
 using System.Text;
 
-namespace Lombiq.Tests.UI.Services;
+namespace Lombiq.Tests.UI.AccessibilityChecking;
 
 /// <summary>
 /// Findings to include in the generated HTML report.
@@ -61,7 +62,7 @@ public enum AxeReportTypes
 /// </para>
 /// <para>
 /// The legacy <c>TWP.Selenium.Axe.Html</c> API surface is preserved through a compatibility shim, see
-/// <c>Lombiq.Tests.UI/Services/AxeHtmlReportShim.cs</c>.
+/// <c>Lombiq.Tests.UI/AccessibilityChecking/AxeHtmlReportShim.cs</c>.
 /// </para>
 /// </remarks>
 public static class AxeHtmlReport
@@ -425,207 +426,17 @@ public static class AxeHtmlReport
 
     private static string EncodeAttribute(string? value) => WebUtility.HtmlEncode(value ?? string.Empty);
 
-    private const string Css = @"
-.thumbnail {
-    content: url('{{SCREENSHOT_URL}}');
-    border: 1px solid black;
-    margin-left: 1em;
-    margin-right: 1em;
-    width: auto;
-    max-height: 150px;
-}
+    private static readonly string Css = LoadEmbeddedResource("AxeHtmlReport.css");
 
-.thumbnail:hover {
-    border: 2px solid black;
-}
-
-.wrapOne,
-.wrapTwo,
-.wrapThree {
-    margin: 2px;
-    max-width: 70vw;
-}
-
-.wrapOne {
-    margin-left: 1em;
-}
-
-.wrapTwo {
-    margin-left: 2em;
-}
-
-.wrapThree {
-    margin-left: 3em;
-}
-
-.emOne {
-    margin-left: 1em;
-    margin-right: 1em;
-}
-
-.emTwo {
-    margin-left: 2em;
-}
-
-.emThree {
-    margin-left: 3em;
-}
-
-#modal {
-    display: none;
-    position: fixed;
-    z-index: 1;
-    left: 0;
-    top: 0;
-    width: 100%;
-    height: 100%;
-    overflow: auto;
-    background-color: rgba(0, 0, 0, 0.9);
-    flex-direction: column;
-}
-
-#modalclose {
-    font-family: Lucida Console, sans-serif;
-    font-size: 35px;
-    width: auto;
-    color: white;
-    text-align: right;
-    padding: 20px;
-    cursor: pointer;
-    max-height: 10%;
-}
-
-#modalimage {
-    margin: auto;
-    display: block;
-    max-width: 95%;
-    padding: 10px;
-    max-height: 90%;
-}
-
-.htmlTable {
-    border-top: double lightgray;
-    width: 100%;
-    display: table;
-}
-
-.sectionbutton {
-    background-color: #000000;
-    color: #ffffff;
-    cursor: pointer;
-    padding: 18px;
-    width: 100%;
-    text-align: left;
-    outline: none;
-    transition: 0.4s;
-    border: 1px solid black;
-}
-
-.sectionbutton:hover {
-    background-color: #828282;
-}
-
-.buttonInfoText {
-    width: 50%;
-    float: left;
-}
-
-.buttonExpandoText {
-    text-align: right;
-    width: 50%;
-    float: right;
-}
-
-.majorSection {
-    padding: 0 18px;
-    background-color: white;
-    overflow: hidden;
-    transition: max-height 0.2s ease-out;
-}
-
-.findings {
-    margin-top: 5px;
-    border-top: 1px solid black;
-}
-
-.active {
-    background-color: #474747;
-    margin-bottom: 0;
-}
-
-.resultWrapper {
-    margin: 5px;
-}
-
-#context {
-    width: 50%;
-}
-
-#image {
-    width: 50%;
-    height: 220px;
-}
-
-#counts {
-    width: 100%;
-}
-
-#metadata {
-    display: flex;
-    flex-wrap: wrap;
-}
-
-#results {
-    display: flex;
-    flex-direction: column;
-}
-
-@media only screen and (max-width: 800px) {
-    #metadata {
-        flex-direction: column;
+    private static string LoadEmbeddedResource(string name)
+    {
+        using var stream = typeof(AxeHtmlReport).Assembly
+            .GetManifestResourceStream($"Lombiq.Tests.UI.AccessibilityChecking.{name}");
+        if (stream is null) throw new InvalidOperationException($"Embedded resource '{name}' not found.");
+        using var reader = new StreamReader(stream);
+        return reader.ReadToEnd();
     }
 
-    #context,
-    #image {
-        width: 100%;
-    }
-}";
-
-    private const string Js = @"
-const buttons = document.getElementsByClassName('sectionbutton');
-
-for (let i = 0; i < buttons.length; i++) {
-    buttons[i].addEventListener('click', function () {
-        const expandoText = this.getElementsByClassName('buttonExpandoText')[0];
-        this.classList.toggle('active');
-
-        const content = this.nextElementSibling;
-        if (expandoText.innerHTML === '-') {
-            content.style.maxHeight = '0';
-            expandoText.innerHTML = '+';
-        }
-        else {
-            content.style.maxHeight = content.scrollHeight + 'px';
-            expandoText.innerHTML = '-';
-        }
-    });
+    private static readonly string Js = LoadEmbeddedResource("AxeHtmlReport.js");
 }
-
-const thumbnail = document.getElementById('screenshotThumbnail');
-const modal = document.getElementById('modal');
-const modalImage = document.getElementById('modalimage');
-
-if (thumbnail && modal && modalImage) {
-    thumbnail.addEventListener('click', function () {
-        modal.style.display = 'flex';
-        modalImage.src = thumbnail.currentSrc || thumbnail.src;
-        modalImage.alt = thumbnail.alt;
-    });
-
-    modal.addEventListener('click', function () {
-        modal.style.display = 'none';
-        modalImage.src = '';
-        modalImage.alt = '';
-    });
-}";
-}
+
