+++ Lombiq.Tests.UI/AccessibilityChecking/AxeHtmlReport.css
@@ -1,0 +1,164 @@
+.thumbnail {
+    content: url('{{SCREENSHOT_URL}}');
+    border: 1px solid black;
+    margin-left: 1em;
+    margin-right: 1em;
+    width: auto;
+    max-height: 150px;
+}
+
+.thumbnail:hover {
+    border: 2px solid black;
+}
+
+.wrapOne,
+.wrapTwo,
+.wrapThree {
+    margin: 2px;
+    max-width: 70vw;
+}
+
+.wrapOne {
+    margin-left: 1em;
+}
+
+.wrapTwo {
+    margin-left: 2em;
+}
+
+.wrapThree {
+    margin-left: 3em;
+}
+
+.emOne {
+    margin-left: 1em;
+    margin-right: 1em;
+}
+
+.emTwo {
+    margin-left: 2em;
+}
+
+.emThree {
+    margin-left: 3em;
+}
+
+#modal {
+    display: none;
+    position: fixed;
+    z-index: 1;
+    left: 0;
+    top: 0;
+    width: 100%;
+    height: 100%;
+    overflow: auto;
+    background-color: rgba(0, 0, 0, 0.9);
+    flex-direction: column;
+}
+
+#modalclose {
+    font-family: Lucida Console, sans-serif;
+    font-size: 35px;
+    width: auto;
+    color: white;
+    text-align: right;
+    padding: 20px;
+    cursor: pointer;
+    max-height: 10%;
+}
+
+#modalimage {
+    margin: auto;
+    display: block;
+    max-width: 95%;
+    padding: 10px;
+    max-height: 90%;
+}
+
+.htmlTable {
+    border-top: double lightgray;
+    width: 100%;
+    display: table;
+}
+
+.sectionbutton {
+    background-color: #000000;
+    color: #ffffff;
+    cursor: pointer;
+    padding: 18px;
+    width: 100%;
+    text-align: left;
+    outline: none;
+    transition: 0.4s;
+    border: 1px solid black;
+}
+
+.sectionbutton:hover {
+    background-color: #828282;
+}
+
+.buttonInfoText {
+    width: 50%;
+    float: left;
+}
+
+.buttonExpandoText {
+    text-align: right;
+    width: 50%;
+    float: right;
+}
+
+.majorSection {
+    padding: 0 18px;
+    background-color: white;
+    overflow: hidden;
+    transition: max-height 0.2s ease-out;
+}
+
+.findings {
+    margin-top: 5px;
+    border-top: 1px solid black;
+}
+
+.active {
+    background-color: #474747;
+    margin-bottom: 0;
+}
+
+.resultWrapper {
+    margin: 5px;
+}
+
+#context {
+    width: 50%;
+}
+
+#image {
+    width: 50%;
+    height: 220px;
+}
+
+#counts {
+    width: 100%;
+}
+
+#metadata {
+    display: flex;
+    flex-wrap: wrap;
+}
+
+#results {
+    display: flex;
+    flex-direction: column;
+}
+
+@media only screen and (max-width: 800px) {
+    #metadata {
+        flex-direction: column;
+    }
+
+    #context,
+    #image {
+        width: 100%;
+    }
+}
