+++ Lombiq.Tests.UI/Services/UITestExecutionSession.cs
@@ -1,5 +1,6 @@
 using Atata.HtmlValidation;
 using Cysharp.Text;
+using Lombiq.Tests.UI.AccessibilityChecking;
 using Lombiq.HelpfulLibraries.Common.Utilities;
 using Lombiq.Tests.UI.Constants;
 using Lombiq.Tests.UI.Exceptions;
